cpp: bioformats: Split common structures from OMETIFFReader into OMETIFF.h
To be shared with the OMETIFFWriter
