+++ cpp/lib/ome/bioformats/CMakeLists.txt
@@ -88,7 +88,8 @@ set(OME_BIOFORMATS_DETAIL_SOURCES
 
 set(OME_BIOFORMATS_DETAIL_HEADERS
     detail/FormatReader.h
-    detail/FormatWriter.h)
+    detail/FormatWriter.h
+    detail/OMETIFF.h)
 
 set(OME_BIOFORMATS_IN_SOURCES
     in/MinimalTIFFReader.cpp
